+++ apps/api/src/test/kotlin/dev/sumin/skeleton/TestcontainersConfiguration.kt
@@ -12,7 +12,7 @@ class TestcontainersConfiguration {
 	@Bean
 	@ServiceConnection
 	fun mysqlContainer(): MySQLContainer {
-		return MySQLContainer(DockerImageName.parse("mysql:latest"))
+		return MySQLContainer(DockerImageName.parse("mysql:8.4"))
 	}
 
 }